+++ stockpot/app/controllers/stockpot/application_controller.rb
@@ -1,0 +1,5 @@
+module Stockpot
+  class ApplicationController < ActionController::API
+    # protect_from_forgery with: :exception
+  end
+end