+++ stockpot/app/controllers/stockpot/database_management_controller.rb
@@ -1,0 +1,9 @@
+require_dependency "stockpot/application_controller"
+
+module Stockpot
+  class DatabaseManagementController < ApplicationController
+    def index
+      render json: {"stockpot": "hello"}.to_json, status: :ok
+    end
+  end
+end