+++ stockpot/app/jobs/stockpot/application_job.rb
@@ -1,0 +1,4 @@
+module Stockpot
+  class ApplicationJob < ActiveJob::Base
+  end
+end