+++ stockpot/app/models/stockpot/application_record.rb
@@ -1,0 +1,5 @@
+module Stockpot
+  class ApplicationRecord < ActiveRecord::Base
+    self.abstract_class = true
+  end
+end