+++ stockpot/config/routes.rb
@@ -1,0 +1,3 @@
+Stockpot::Engine.routes.draw do
+  get '/', to: "database_management#index"
+end