+++ stockpot/lib/stockpot/engine.rb
@@ -1,0 +1,6 @@
+module Stockpot
+  class Engine < ::Rails::Engine
+    isolate_namespace Stockpot
+    config.generators.api_only = true
+  end
+end