+++ stockpot/spec/spec_helper.rb
@@ -2,6 +2,4 @@
 
 require_relative "../../shared/spec_helper"
 
-require_relative "support/shared_context/with_an_example_collection"
-
 require "stockpot"